# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Learn Japanese API Documentation",
        "description": "API endpoints for Learn Japanese application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8081/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/vocabularies": {
            "get": {
                "tags": [
                    "Vocabularies"
                ],
                "summary": "List Vocabularies",
                "description": "Display a listing of vocabularies, with optional search, filtering, and pagination.",
                "operationId": "5900ce3deeaae2bf9ce6fcfa85d72f89",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "paginate",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "true"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/vocabularies/{id}": {
            "get": {
                "tags": [
                    "Vocabularies"
                ],
                "summary": "Vocabulary Details",
                "description": "Display the specified vocabulary item with related group and kanjis.",
                "operationId": "50c7733471db90e14118073cd504ac44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/vocabularies/random": {
            "get": {
                "tags": [
                    "Vocabularies"
                ],
                "summary": "Random Flashcards",
                "description": "Get a list ramdom vocabularies of flashcards",
                "operationId": "18425105f42fb392f9ad97d503893824",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Vocabulary"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Vocabulary": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "group_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "word": {
                        "type": "string",
                        "example": "一つ"
                    },
                    "furigana": {
                        "type": "string",
                        "example": "ひとつ",
                        "nullable": true
                    },
                    "romaji": {
                        "type": "string",
                        "example": "hitotsu"
                    },
                    "meaning": {
                        "type": "string",
                        "example": "Một cái"
                    },
                    "type": {
                        "type": "string",
                        "example": "numbers_counters",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}